solar_project/requirements.txt
Django>=4.2
djangorestframework>=3.14
pandas>=1.5
numpy>=1.23
scikit-learn>=1.2
joblib>=1.2
requests>=2.28
drf-yasg>=1.21
requests
pandas
scikit-learn==1.5.0
